+++ test/data/test_ecmwf.py
@@ -73,7 +73,7 @@ def test_ifs_fetch(time, variable):
     assert shape[3] == 721
     assert shape[4] == 1440
     assert np.array_equal(data.coords["variable"].values, np.array(variable))
-    assert np.all_close(data.values, data_fx.values[:, 0])
+    assert np.allclose(data.values, data_fx.values[:, 0])
 
 
 @pytest.mark.slow
@@ -90,9 +90,9 @@ def test_ifs_fetch(time, variable):
     ],
 )
 @pytest.mark.parametrize("variable", ["tcwv", ["sp"]])
-def test_ifs_ens_fetch(time, variable):
-    members = [1, 2, 3]
-    ds = IFS_ENS(cache=False, members=members)
+@pytest.mark.parametrize("member", [0, 1])
+def test_ifs_ens_fetch(time, variable, member):
+    ds = IFS_ENS(cache=False, member=member)
     data = ds(time, variable)
     shape = data.shape
 
@@ -104,24 +104,22 @@ def test_ifs_ens_fetch(time, variable):
 
     assert shape[0] == len(time)
     assert shape[1] == len(variable)
-    assert shape[2] == len(members)
-    assert shape[3] == 721
-    assert shape[4] == 1440
+    assert shape[2] == 721
+    assert shape[3] == 1440
     assert not np.isnan(data.values).any()
     assert np.array_equal(data.coords["variable"].values, np.array(variable))
 
-    ds = IFS_ENS_FX(cache=False, members=members)
+    ds = IFS_ENS_FX(cache=False, member=member)
     data_fx = ds(time, timedelta(hours=0), variable)
     shape = data_fx.shape
 
     assert shape[0] == len(time)
     assert shape[1] == 1
     assert shape[2] == len(variable)
-    assert shape[3] == len(members)
-    assert shape[4] == 721
-    assert shape[5] == 1440
+    assert shape[3] == 721
+    assert shape[4] == 1440
     assert not np.isnan(data.values).any()
-    assert np.all_close(data.values, data_fx.values[:, 0])
+    assert np.allclose(data.values, data_fx.values[:, 0])
 
 
 @pytest.mark.slow
@@ -149,8 +147,7 @@ def test_ifs_ens_fetch(time, variable):
 )
 @pytest.mark.parametrize("variable", ["tcwv", ["sp"]])
 def test_ifs_ens_fx_fetch(time, lead_time, variable):
-    members = [1, 2, 3]
-    ds = IFS_ENS_FX(cache=False, members=members)
+    ds = IFS_ENS_FX(cache=False, member=0)
     data = ds(time, lead_time, variable)
     shape = data.shape
 
@@ -166,9 +163,47 @@ def test_ifs_ens_fx_fetch(time, lead_time, variable):
     assert shape[0] == len(time)
     assert shape[1] == len(lead_time)
     assert shape[2] == len(variable)
-    assert shape[3] == len(members)
-    assert shape[4] == 721
-    assert shape[5] == 1440
+    assert shape[3] == 721
+    assert shape[4] == 1440
+    assert not np.isnan(data.values).any()
+    assert np.array_equal(data.coords["variable"].values, np.array(variable))
+
+
+@pytest.mark.slow
+@pytest.mark.xfail
+@pytest.mark.timeout(30)
+@pytest.mark.parametrize(
+    "lead_time",
+    [
+        timedelta(hours=0),
+        [
+            timedelta(hours=6),
+            timedelta(hours=12),
+        ],
+    ],
+)
+@pytest.mark.parametrize("variable", ["t2m", ["sp"]])
+@pytest.mark.parametrize("member", [0, 1])
+def test_aifs_ens_fx_fetch(lead_time, variable, member):
+    time = now6h() - timedelta(hours=12)
+    ds = AIFS_ENS_FX(cache=False, member=member)
+    data = ds(time, lead_time, variable)
+    shape = data.shape
+
+    if isinstance(variable, str):
+        variable = [variable]
+
+    if isinstance(lead_time, timedelta):
+        lead_time = [lead_time]
+
+    if isinstance(time, datetime):
+        time = [time]
+
+    assert shape[0] == len(time)
+    assert shape[1] == len(lead_time)
+    assert shape[2] == len(variable)
+    assert shape[3] == 721
+    assert shape[4] == 1440
     assert not np.isnan(data.values).any()
     assert np.array_equal(data.coords["variable"].values, np.array(variable))
 
@@ -222,11 +257,8 @@ def test_ifs_cache(time, variable, cache):
         datetime(year=1993, month=4, day=5),
     ],
 )
-@pytest.mark.parametrize(
-    "lead_time",
-)
 @pytest.mark.parametrize("variable", ["msl"])
-def test_ifs_time_available(data_source, time, lead_time, variable):
+def test_ifs_time_available(data_source, time, variable):
     with pytest.raises(ValueError):
         ds = IFS(source="ecmwf")
         ds(time, ["msl"])
@@ -236,8 +268,6 @@ def test_ifs_time_available(data_source, time, lead_time, variable):
 
 
 @pytest.mark.timeout(30)
-@pytest.mark.parametrize("data_source", [IFS, IFS_ENS])
-@pytest.mark.parametrize("time", [now6h() - timedelta(days=2)])
 @pytest.mark.parametrize(
     "lead_time",
     [
@@ -246,11 +276,14 @@ def test_ifs_time_available(data_source, time, lead_time, variable):
         timedelta(hours=366),
     ],
 )
-@pytest.mark.parametrize("variable", ["msl"])
-def test_ifs_leadtime_available(data_source, time, lead_time, variable):
+def test_ifs_leadtime_available(lead_time):
     with pytest.raises(ValueError):
-        ds = data_source()
-        ds(time, lead_time, variable)
+        time = now6h() - timedelta(days=2)
+        ds = IFS_FX(source="ecmwf")
+        ds(time, lead_time, ["msl"])
+
+        ds = IFS_ENS_FX(source="ecmwf")
+        ds(time, lead_time, ["msl"])
 
 
 @pytest.mark.timeout(30)